review, image added for 2pointers, etc
