--- _includes/source_code/code/33-PDP/landfight/landfight.in
@@ -1,20 +1,0 @@
-7
-3 2 5 7 1 6 4
-//2
-
-9
-1    2    1    3   1  1 1   7  1 
-1    3    4    7   8  9 10 17 18 PS
-18  17   15   14  11 10 9   8  1 SS
-
-7
-6 -1 -1 -1 -1 -1 -1
-//0
-
-
-10
-5 -2 3 -7 -4 -1 3 -5 6 0
-//0
-
-
-
