--- _includes/source_code/code/33-PDP/landfight/landfight.out
@@ -1,1 +1,0 @@
-2
