+++ _includes/source_code/code/33-PDP/landfight/landfight_vector.cc
@@ -7,25 +7,7 @@ using namespace std;
 const long MAXN = 1'000'000L;
 
 long A[MAXN];
-struct myvector {
-    long head;
-    vector<long> pos;
-    myvector(){
-        head = -1;//δεν υπάρχουν στοιχεία στον πίνακα
-    }
-    long query(long i){
-        //αφαίρεσε τις θέσεις που ξεπεράστηκαν
-        while(head>=0 && pos[head]<=i)
-            head--;
-        return (head>=0)?pos[head]:(MAXN+1);
-    }
-    void push(long x){
-        pos.push_back(x);
-        head++;//δείκτης στο τελευταίο στοιχείο
-    }
-};
-
-unordered_map<int, myvector> M;
+unordered_map<int, vector<long> > M;
 
 int main(){
     freopen("landfight.in","r",stdin);
@@ -36,18 +18,23 @@ int main(){
     for(long i=1;i<=N;i++)
         scanf("%ld",A+i);
     
-    M[0].push(N+1);//περίπτωση L=N,R=0
+    M[0].push_back(N+1);//περίπτωση L=N,R=0
     for(long suffix=0L,i=N;i>0;i--){
         suffix +=A[i];
-        M[suffix].push(i);
+        M[suffix].push_back(i);
     }
     
     long ans = N;
     for(long prefix=0L,i=1;i<=N;i++){
         prefix += A[i];
         if(M.find(prefix) == M.end())
-            continue;        
-        ans = min(ans,M[prefix].query(i)-i-1);
+            continue;
+        vector<long>& V = M[prefix];
+        //διέγραψε τα ξεπερασμένα
+        while(!V.empty() && V.back()<=i)
+            V.pop_back();
+        if(!V.empty())
+            ans = min(ans,V.back()-i-1);
     }
     
     printf("%ld\n",ans);
