+++ contests/_33-PDP/c-landfight-statement.md
@@ -27,7 +27,7 @@ $$x_1,x_2, \ldots x_N$$. Μερικά χωράφια βγάζουν κέρδος
 
 Το αρχείο εξόδου με όνομα **landfight.out** είναι αρχείο κειμένου που περιέχει μία μόνο γραμμή με έναν ακέραιο αριθμό: την απάντηση στο παραπάνω ερώτημα.
 
-## Παραδείγμα Αρχείων Εισόδου - Εξόδου:
+## Παραδείγματα Αρχείων Εισόδου - Εξόδου:
 
 ## 1<sup>o</sup>
 
@@ -65,10 +65,9 @@ $$x_1,x_2, \ldots x_N$$. Μερικά χωράφια βγάζουν κέρδος
  * Για περιπτώσεις ελέγχου συνολικής αξίας 70%, θα είναι $$x_i \gt 0$$
  
 ## Προσοχή:
+
 Φροντίστε να διαβάζετε την είσοδο και να εκτυπώνετε την έξοδο αποδοτικά, ειδικά αν προγραμματίζετε σε C++ ή Java.
 
-**Μορφοποίηση**: Στην έξοδο, όλες οι γραμμές τερματίζουν με έναν χαρακτήρα newline.
-
-**Μέγιστος χρόνος εκτέλεσης**: $$1$$ sec.
-
-**Μέγιστη διαθέσιμη μνήμη**: $$64$$ MB.
+**Μορφοποίηση**: Στην έξοδο, όλες οι γραμμές τερματίζουν με έναν χαρακτήρα newline.<br>
+**Μέγιστος χρόνος εκτέλεσης**: $$1$$ sec.<br>
+**Μέγιστη διαθέσιμη μνήμη**: $$64$$ MB.<br>
